int to double pour la surface

+++ src/main/java/com/example/geodevineur/dep_reg/Departement.java
@@ -4,7 +4,7 @@ import com.example.geodevineur.enumerations.*;
 
 public class Departement extends DepReg {
 
-    public Departement (String name_, int pop_, int surf_, Cardinal card_, boolean sea_, int neigh_, String id_, Politic politic_) {
+    public Departement (String name_, int pop_, double surf_, Cardinal card_, boolean sea_, int neigh_, String id_, Politic politic_) {
         name = name_;
         population = pop_;
         surface = surf_;
